// playground/types/supabase.ts
export interface definitions {
  /** Full list of countries. */
  countries: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    /** Full country name. */
    name?: string;
    /** ISO 3166-1 alpha-2 code. */
    iso2: string;
    /** ISO 3166-1 alpha-3 code. */
    iso3?: string;
    /** Local variation of the name. */
    local_name?: string;
    continent?:
      | "Africa"
      | "Antarctica"
      | "Asia"
      | "Europe"
      | "Oceania"
      | "North America"
      | "South America";
  };
}